+++ pkg/providers/cloudcertificates/testdata/TestDataCloudCertificatesHostnameBindings/single.tf
@@ -1,0 +1,6 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_cloudcertificates_hostname_bindings" "test" {
+}